+++ examples/0_basic/5_create_vc.rs
@@ -11,7 +11,9 @@
 
 use examples::create_did;
 use examples::MemStorage;
+use identity_eddsa_verifier::EdDSAJwsVerifier;
 use identity_iota::core::Object;
+
 use identity_iota::credential::DecodedJwtCredential;
 use identity_iota::credential::Jwt;
 use identity_iota::credential::JwtCredentialValidationOptions;
@@ -86,7 +88,13 @@ async fn main() -> anyhow::Result<()> {
     .build()?;
 
   let credential_jwt: Jwt = issuer_document
-    .create_credential_jwt(&credential, &issuer_storage, &fragment, &JwsSignatureOptions::default())
+    .create_credential_jwt(
+      &credential,
+      &issuer_storage,
+      &fragment,
+      &JwsSignatureOptions::default(),
+      None,
+    )
     .await?;
 
   // Before sending this credential to the holder the issuer wants to validate that some properties
@@ -94,14 +102,15 @@ async fn main() -> anyhow::Result<()> {
 
   // Validate the credential's signature using the issuer's DID Document, the credential's semantic structure,
   // that the issuance date is not in the future and that the expiration date is not in the past:
-  let decoded_credential: DecodedJwtCredential<Object> = JwtCredentialValidator::new()
-    .validate::<_, Object>(
-      &credential_jwt,
-      &issuer_document,
-      &JwtCredentialValidationOptions::default(),
-      FailFast::FirstError,
-    )
-    .unwrap();
+  let decoded_credential: DecodedJwtCredential<Object> =
+    JwtCredentialValidator::with_signature_verifier(EdDSAJwsVerifier::default())
+      .validate::<_, Object>(
+        &credential_jwt,
+        &issuer_document,
+        &JwtCredentialValidationOptions::default(),
+        FailFast::FirstError,
+      )
+      .unwrap();
 
   println!("VC successfully validated");