+++ identity_credential/src/error.rs
@@ -45,6 +45,11 @@ pub enum Error {
   #[error("could not convert JWT to the VC data model: {0}")]
   InconsistentCredentialJwtClaims(&'static str),
 
+  /// Caused when deserializing a Presentation with an empty array for the
+  /// `verifiableCredential` property.
+  #[error("empty verifiableCredential array in presentation")]
+  EmptyVerifiableCredentialArray,
+
   /// Caused when attempting to convert a JWT to a `Presentation` that has conflicting values
   /// between the registered claims and those in the `vp` object.
   #[error("could not convert JWT to the VP data model: {0}")]