+++ identity_storage/src/storage/jwk_document_ext.rs
@@ -15,6 +15,7 @@ use super::JwsSignatureOptions;
 use super::Storage;
 
 use async_trait::async_trait;
+use identity_core::common::Object;
 use identity_credential::credential::Credential;
 use identity_credential::credential::Jws;
 use identity_credential::credential::Jwt;
@@ -95,16 +96,20 @@ pub trait JwkDocumentExt: private::Sealed {
     I: KeyIdStorage;
 
   /// Produces a JWT where the payload is produced from the given `credential`
-  /// in accordance with [VC-JWT version 1.1](https://w3c.github.io/vc-jwt/#version-1.1).
+  /// in accordance with [VC Data Model v1.1](https://www.w3.org/TR/vc-data-model/#json-web-token).
   ///
-  /// The `kid` in the protected header is the `id` of the method identified by `fragment` and the JWS signature will be
-  /// produced by the corresponding private key backed by the `storage` in accordance with the passed `options`.
+  /// Unless the `kid` is explicitly set in the options, the `kid` in the protected header is the `id`
+  /// of the method identified by `fragment` and the JWS signature will be produced by the corresponding
+  /// private key backed by the `storage` in accordance with the passed `options`.
+  ///
+  /// The `custom_claims` can be used to set additional claims on the resulting JWT.
   async fn create_credential_jwt<K, I, T>(
     &self,
     credential: &Credential<T>,
     storage: &Storage<K, I>,
     fragment: &str,
     options: &JwsSignatureOptions,
+    custom_claims: Option<Object>,
   ) -> StorageResult<Jwt>
   where
     K: JwkStorage,
@@ -112,10 +117,11 @@ pub trait JwkDocumentExt: private::Sealed {
     T: ToOwned<Owned = T> + Serialize + DeserializeOwned + Sync;
 
   /// Produces a JWT where the payload is produced from the given `presentation`
-  /// in accordance with [VC-JWT version 1.1](https://w3c.github.io/vc-jwt/#version-1.1).
+  /// in accordance with [VC Data Model v1.1](https://www.w3.org/TR/vc-data-model/#json-web-token).
   ///
-  /// The `kid` in the protected header is the `id` of the method identified by `fragment` and the JWS signature will be
-  /// produced by the corresponding private key backed by the `storage` in accordance with the passed `options`.
+  /// Unless the `kid` is explicitly set in the options, the `kid` in the protected header is the `id`
+  /// of the method identified by `fragment` and the JWS signature will be produced by the corresponding
+  /// private key backed by the `storage` in accordance with the passed `options`.
   async fn create_presentation_jwt<K, I, CRED, T>(
     &self,
     presentation: &Presentation<CRED, T>,
@@ -356,8 +362,15 @@ impl JwkDocumentExt for CoreDocument {
       let mut header = JwsHeader::new();
 
       header.set_alg(alg);
+      if let Some(custom) = &options.custom_header_parameters {
+        header.set_custom(custom.clone())
+      }
 
-      header.set_kid(method.id().to_string());
+      if let Some(ref kid) = options.kid {
+        header.set_kid(kid.clone());
+      } else {
+        header.set_kid(method.id().to_string());
+      }
 
       if options.attach_jwk {
         header.set_jwk(jwk.clone())
@@ -373,7 +386,10 @@ impl JwkDocumentExt for CoreDocument {
 
       if let Some(typ) = &options.typ {
         header.set_typ(typ.clone())
-      };
+      } else {
+        // https://www.w3.org/TR/vc-data-model/#jwt-encoding
+        header.set_typ("JWT")
+      }
 
       if let Some(cty) = &options.cty {
         header.set_cty(cty.clone())
@@ -386,6 +402,7 @@ impl JwkDocumentExt for CoreDocument {
       if let Some(nonce) = &options.nonce {
         header.set_nonce(nonce.clone())
       };
+
       header
     };
 
@@ -420,6 +437,7 @@ impl JwkDocumentExt for CoreDocument {
     storage: &Storage<K, I>,
     fragment: &str,
     options: &JwsSignatureOptions,
+    custom_claims: Option<Object>,
   ) -> StorageResult<Jwt>
   where
     K: JwkStorage,
@@ -439,7 +457,9 @@ impl JwkDocumentExt for CoreDocument {
       )));
     }
 
-    let payload = credential.serialize_jwt().map_err(Error::ClaimsSerializationError)?;
+    let payload = credential
+      .serialize_jwt(custom_claims)
+      .map_err(Error::ClaimsSerializationError)?;
     self
       .create_jws(storage, fragment, payload.as_bytes(), options)
       .await
@@ -563,6 +583,7 @@ mod iota_document {
       storage: &Storage<K, I>,
       fragment: &str,
       options: &JwsSignatureOptions,
+      custom_claims: Option<Object>,
     ) -> StorageResult<Jwt>
     where
       K: JwkStorage,
@@ -571,7 +592,7 @@ mod iota_document {
     {
       self
         .core_document()
-        .create_credential_jwt(credential, storage, fragment, options)
+        .create_credential_jwt(credential, storage, fragment, options, custom_claims)
         .await
     }
     async fn create_presentation_jwt<K, I, CRED, T>(