+++ identity_storage/src/storage/tests/credential_jws.rs
@@ -6,8 +6,10 @@ use identity_core::convert::FromJson;
 use identity_credential::credential::Credential;
 
 use identity_credential::validator::JwtCredentialValidationOptions;
+use identity_did::DID;
 use identity_document::document::CoreDocument;
 use identity_document::verifiable::JwsVerificationOptions;
+use identity_eddsa_verifier::EdDSAJwsVerifier;
 use identity_verification::jose::jws::JwsAlgorithm;
 use identity_verification::MethodScope;
 
@@ -83,6 +85,7 @@ async fn signing_credential_with_detached_option_fails() {
       &storage,
       kid.as_ref(),
       &JwsSignatureOptions::default().detached_payload(true),
+      None
     )
     .await
     .is_err());
@@ -99,11 +102,13 @@ async fn signing_credential_with_nonce_and_scope() {
       &storage,
       kid.as_ref(),
       &JwsSignatureOptions::default().nonce(nonce.to_owned()),
+      None,
     )
     .await
     .unwrap();
 
-  let validator = identity_credential::validator::JwtCredentialValidator::new();
+  let validator =
+    identity_credential::validator::JwtCredentialValidator::with_signature_verifier(EdDSAJwsVerifier::default());
   assert!(validator
     .validate::<_, Object>(
       &jws,
@@ -156,11 +161,13 @@ async fn signing_credential_with_b64() {
       &storage,
       kid.as_ref(),
       &JwsSignatureOptions::default().b64(true),
+      None,
     )
     .await
     .unwrap();
 
-  let validator = identity_credential::validator::JwtCredentialValidator::new();
+  let validator =
+    identity_credential::validator::JwtCredentialValidator::with_signature_verifier(EdDSAJwsVerifier::default());
   let decoded = validator
     .validate::<_, Object>(
       &jws,
@@ -182,7 +189,113 @@ async fn signing_credential_with_b64() {
       &storage,
       kid.as_ref(),
       &JwsSignatureOptions::default().b64(false),
+      None
     )
     .await
     .is_err());
+}
+
+#[tokio::test]
+async fn signing_credential_with_custom_kid() {
+  let (document, storage, fragment, credential) = setup().await;
+
+  let my_kid = "my-kid";
+  let jws = document
+    .create_credential_jwt(
+      &credential,
+      &storage,
+      fragment.as_ref(),
+      &JwsSignatureOptions::default().kid(my_kid),
+      None,
+    )
+    .await
+    .unwrap();
+
+  let validator =
+    identity_credential::validator::JwtCredentialValidator::with_signature_verifier(EdDSAJwsVerifier::default());
+  let method_id = document.id().clone().join(format!("#{fragment}")).unwrap();
+  let decoded = validator
+    .validate::<_, Object>(
+      &jws,
+      &document,
+      &JwtCredentialValidationOptions::default()
+        .verification_options(JwsVerificationOptions::new().method_id(method_id)),
+      identity_credential::validator::FailFast::FirstError,
+    )
+    .unwrap();
+
+  assert_eq!(decoded.header.kid().unwrap(), my_kid);
+}
+
+#[tokio::test]
+async fn custom_claims() {
+  let (document, storage, kid, credential) = setup().await;
+
+  let mut custom_claims = Object::new();
+  custom_claims.insert(
+    "test-key".to_owned(),
+    serde_json::Value::String("test-value".to_owned()),
+  );
+  let jws = document
+    .create_credential_jwt(
+      &credential,
+      &storage,
+      kid.as_ref(),
+      &JwsSignatureOptions::default().b64(true),
+      Some(custom_claims.clone()),
+    )
+    .await
+    .unwrap();
+
+  let validator =
+    identity_credential::validator::JwtCredentialValidator::with_signature_verifier(EdDSAJwsVerifier::default());
+  let decoded = validator
+    .validate::<_, Object>(
+      &jws,
+      &document,
+      &JwtCredentialValidationOptions::default(),
+      identity_credential::validator::FailFast::FirstError,
+    )
+    .unwrap();
+  assert_eq!(decoded.custom_claims.unwrap(), custom_claims);
+}
+
+#[tokio::test]
+async fn custom_header_parameters() {
+  let (document, storage, kid, credential) = setup().await;
+
+  let mut custom = Object::new();
+  custom.insert(
+    "test-key".to_owned(),
+    serde_json::Value::String("test-value".to_owned()),
+  );
+  let jws = document
+    .create_credential_jwt(
+      &credential,
+      &storage,
+      kid.as_ref(),
+      &JwsSignatureOptions::default()
+        .b64(true)
+        .custom_header_parameters(custom),
+      None,
+    )
+    .await
+    .unwrap();
+
+  let validator =
+    identity_credential::validator::JwtCredentialValidator::with_signature_verifier(EdDSAJwsVerifier::default());
+  let decoded = validator
+    .validate::<_, Object>(
+      &jws,
+      &document,
+      &JwtCredentialValidationOptions::default(),
+      identity_credential::validator::FailFast::FirstError,
+    )
+    .unwrap();
+  let custom_from_decoded = decoded.header.as_ref().custom().unwrap();
+  assert_eq!(custom_from_decoded.len(), 1);
+  assert_eq!(
+    custom_from_decoded.get("test-key").unwrap().as_str().unwrap(),
+    "test-value".to_owned()
+  );
 }